requirements.txt
# Streamlit Community Cloud Deployment Requirements
# Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# NLP Libraries
nltk==3.8.1
spacy==3.6.1

# Deep Learning & Transformers
torch==2.0.1
transformers==4.31.0
sentence-transformers==2.2.2

# Vector Database (CPU version for Streamlit Cloud)
faiss-cpu==1.7.4

# API Framework (for local development, not used on Streamlit Cloud)
fastapi==0.100.1
uvicorn[standard]==0.23.2
pydantic==2.1.1

# Web Application
streamlit==1.25.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
tqdm==4.65.0
joblib==1.3.1
requests==2.31.0
pyyaml==6.0.1

# Optional: For hybrid search
rank-bm25==0.2.2
